timer: handle timer overflow
Tock does not support setting alarms more than 2^32 ticks in the future as the
`alarm_at` syscall takes in ticks. For the default nrf52840 tock configuration
at 32KHz, the greatest possible value for the syscall alarm is 37 hours.

`timer_in` takes in milliseconds, but is flawed in that it is just a wrapper
around `alarm_at` which can only accept a max value of 2^32 ticks. Thus,
calling `timer_in` longer than 2^32 ticks in milliseconds will not have the
expected behavior. `delay_ms` calls `timer_in`, so it is flawed for the same
reason.

We fixed this by modifying `timer_in`. If the timer length is longer than the
time that `alarm_at` can count up to, then `timer_in` will schedule multiple
alarms to reach the full length. We calculate the number of full overflows
and the remainder ticks to reach the target length of time. The overflows use
the `overflow_callback` for each intermediate overflow.

Co-Authored-By: Anthony Tarbinian <atar13@users.noreply.github.com>