remove busy waits, replace with yield
Apparently, timer callbacks won't be scheduled if our process is just
busy waiting. We though the kernel would preempt the process to schedule
the callback, but we have learned it does not. Instead, the process must
give up the CPU to the kernel to allow the callback to get run.

Also for some reason we need a print statement after calling yield.
Otherwise the callback does not get scheduled.

Co-Authored-By: Samir Rashid <samir-rashid@users.noreply.github.com>